moved helper function into untTestUtils file
