+++ src/screens/committees/__tests__/committeeUtils.test.ts
@@ -5,135 +5,8 @@ import { Committee } from "../../../types/committees";
 import { checkCommitteeRequestStatus, deleteCommittee, getCommittee, getCommitteeEvents, getCommitteeMembers, getCommittees, getLeads, getPublicUserData, getRepresentatives, getTeamMembers, removeCommitteeRequest, resetCommittee, setCommitteeData, submitCommitteeRequest } from "../../../api/firebaseUtils";
 import { PublicUserInfo, User } from "../../../types/user";
 import { EventType, SHPEEvent } from "../../../types/events";
+import { clearCollection, createTestUserInFirebase, generateTestCommittee, generateTestEvent, generateTestUsers, waitForUser } from "../../../helpers/unitTestUtils";
 
-
-const generateTestCommittee = async (overrides: Partial<Committee> = {}): Promise<Committee> => {
-    return {
-        firebaseDocName: "test-committee",
-        name: "Test Committee",
-        color: "#500000",
-        head: "head1UID",
-        representatives: ["rep1UID", "rep2UID"],
-        leads: ["lead1UID"],
-        applicationLink: "http://testlink.com",
-        logo: "default",
-        memberCount: 10,
-        isOpen: true,
-        ...overrides
-    };
-};
-
-const generateTestEvent = (overrides: Partial<SHPEEvent> = {}): SHPEEvent => {
-    const currentTime = new Date();
-    const startTime = Timestamp.fromDate(currentTime);
-    const endTime = Timestamp.fromDate(new Date(currentTime.getTime() + 3600 * 1000)); // 1 hour in the future
-
-    return {
-        committee: "app-devs",
-        coverImageURI: null,
-        creator: "sampleUID",
-        description: "Test Description",
-        endTime: endTime,
-        endTimeBuffer: 600000,
-        eventType: EventType.INTRAMURAL_EVENT,
-        general: true,
-        geofencingRadius: 100,
-        geolocation: new GeoPoint(30.621160236499136, -96.3403560168198),
-        hiddenEvent: false,
-        locationName: "Test",
-        name: "Test Event",
-        nationalConventionEligible: true,
-        notificationSent: true,
-        signInPoints: 3,
-        startTime: startTime,
-        startTimeBuffer: 600000,
-        ...overrides
-    };
-};
-
-const generateTestUsers = async (overrides: Partial<User> = {}): Promise<User> => {
-    return {
-        publicInfo: {
-            uid: "testUID",
-        },
-        private: {
-            privateInfo: {
-                completedAccountSetup: true,
-                settings: {
-                    darkMode: true,
-                    useSystemDefault: false
-                },
-            },
-        },
-        ...overrides
-    };
-};
-
-const createTestUserInFirebase = async (user: User, maxRetries: number = 20) => {
-    const publicInfo = user.publicInfo;
-    const privateInfo = user.private?.privateInfo;
-    const userDocRef = doc(db, "users", publicInfo?.uid!);
-    const privateInfoDocRef = doc(db, `users/${publicInfo?.uid!}/private`, "privateInfo");
-
-    for (let attempt = 1; attempt <= maxRetries; attempt++) {
-        try {
-            // Attempt to set the documents in Firestore
-            await setDoc(userDocRef, publicInfo);
-            await setDoc(privateInfoDocRef, privateInfo);
-
-            // Verify the creation of the user
-            const testUserDoc = await getDoc(userDocRef);
-            if (testUserDoc.exists()) {
-                return;
-            }
-        } catch (error) {
-            console.error(`Attempt ${attempt} encountered an error: ${error}. Retrying...`);
-        }
-    }
-};
-
-const userExists = async (uid: string) => {
-    const userDocRef = doc(db, "users", uid);
-    const userDoc = await getDoc(userDocRef);
-    return userDoc.exists();
-};
-
-const waitForUser = async (uid: string, maxRetries: number = 16, interval: number = 500, userData?: PublicUserInfo) => {
-    for (let attempt = 1; attempt <= maxRetries; attempt++) {
-        if (await userExists(uid)) {
-            return true;
-        }
-        await createTestUserInFirebase({ publicInfo: userData ? (userData) : ({ uid: uid }) }, 4);
-        await new Promise(resolve => setTimeout(resolve, interval));
-    }
-    throw new Error(`User with UID: ${uid} does not exist after ${maxRetries} attempts`);
-};
-
-
-const clearSubcollections = async (docRef: DocumentReference) => {
-    const subcollectionsSnapshot = await getDocs(collection(docRef, 'private'));
-    const batch = writeBatch(db);
-
-    subcollectionsSnapshot.forEach(subDoc => {
-        batch.delete(subDoc.ref);
-    });
-
-    await batch.commit();
-};
-
-const clearCollection = async (collectionName: string) => {
-    const collectionRef = collection(db, collectionName);
-    const querySnapshot = await getDocs(collectionRef);
-
-    const batch = writeBatch(db);
-
-    for (const documentSnapshot of querySnapshot.docs) {
-        await clearSubcollections(documentSnapshot.ref);
-        batch.delete(documentSnapshot.ref);
-    }
-
-    await batch.commit();
-};
 
 beforeAll(async () => {
     expect(process.env.FIREBASE_EMULATOR_ADDRESS).toBeDefined();
