+++ src/screens/home/__tests__/homeUtils.test.ts
@@ -3,91 +3,7 @@ import { auth, db } from "../../../config/firebaseConfig";
 import { DocumentReference, Timestamp, collection, doc, getDoc, getDocs, setDoc, writeBatch } from "firebase/firestore";
 import { FilterRole, PublicUserInfo, User } from "../../../types/user";
 import { fetchOfficeCount, fetchOfficerStatus, getMOTM, getUserForMemberList, setMOTM, updateOfficerStatus } from "../../../api/firebaseUtils";
-
-
-const generateTestUsers = async (overrides: Partial<User> = {}): Promise<User> => {
-    return {
-        publicInfo: {
-            uid: "testUID",
-        },
-        private: {
-            privateInfo: {
-                completedAccountSetup: true,
-                settings: {
-                    darkMode: true,
-                    useSystemDefault: false
-                },
-            },
-        },
-        ...overrides
-    };
-};
-
-const createTestUserInFirebase = async (user: User, maxRetries: number = 16) => {
-    const publicInfo = user.publicInfo;
-    const privateInfo = user.private?.privateInfo;
-    const userDocRef = doc(db, "users", publicInfo?.uid!);
-    const privateInfoDocRef = doc(db, `users/${publicInfo?.uid!}/private`, "privateInfo");
-
-    for (let attempt = 1; attempt <= maxRetries; attempt++) {
-        try {
-            // Attempt to set the documents in Firestore
-            await setDoc(userDocRef, publicInfo);
-            await setDoc(privateInfoDocRef, privateInfo);
-
-            // Verify the creation of the user
-            const testUserDoc = await getDoc(userDocRef);
-            if (testUserDoc.exists()) {
-                return;
-            }
-        } catch (error) {
-            console.error(`Attempt ${attempt} encountered an error: ${error}. Retrying...`);
-        }
-    }
-};
-
-const userExists = async (uid: string) => {
-    const userDocRef = doc(db, "users", uid);
-    const userDoc = await getDoc(userDocRef);
-    return userDoc.exists();
-};
-
-const waitForUser = async (uid: string, maxRetries: number = 16, interval: number = 500, userData?: PublicUserInfo) => {
-    for (let attempt = 1; attempt <= maxRetries; attempt++) {
-        if (await userExists(uid)) {
-            return true;
-        }
-        await createTestUserInFirebase({ publicInfo: userData ? (userData) : ({ uid: uid }) }, 4);
-        await new Promise(resolve => setTimeout(resolve, interval));
-    }
-    throw new Error(`User with UID: ${uid} does not exist after ${maxRetries} attempts`);
-};
-
-
-const clearSubcollections = async (docRef: DocumentReference) => {
-    const subcollectionsSnapshot = await getDocs(collection(docRef, 'private'));
-    const batch = writeBatch(db);
-
-    subcollectionsSnapshot.forEach(subDoc => {
-        batch.delete(subDoc.ref);
-    });
-
-    await batch.commit();
-};
-
-const clearCollection = async (collectionName: string) => {
-    const collectionRef = collection(db, collectionName);
-    const querySnapshot = await getDocs(collectionRef);
-
-    const batch = writeBatch(db);
-
-    for (const documentSnapshot of querySnapshot.docs) {
-        await clearSubcollections(documentSnapshot.ref);
-        batch.delete(documentSnapshot.ref);
-    }
-
-    await batch.commit();
-};
+import { clearCollection, createTestUserInFirebase, generateTestUsers, waitForUser } from "../../../helpers/unitTestUtils";
 
 beforeAll(async () => {
     expect(process.env.FIREBASE_EMULATOR_ADDRESS).toBeDefined();
